+++ src/main/java/org/kumoricon/site/utility/importattendee/AttendeeFile.java
@@ -1,0 +1,7 @@
+package org.kumoricon.site.utility.importattendee;
+
+import java.util.List;
+
+public class AttendeeFile {
+    public List<AttendeeRecord> attendees;
+}