+++ CHANGES.md
@@ -1,6 +1,8 @@
 Change Log
 ----------------
 - Version 2.0.4 (pending)
+    - Import attendees in JSON format
+    - Added Small Press badge type
 
 - Version 2.0.3 (10/20/2017)
     - Fixed bug that would cause server to crash after running out of files